simplify typing, remove deps
